backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2
gunicorn==21.2.0
numpy==1.26.3
python-dotenv==1.0.0
msal==1.28.0

